Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-dotenv
requests
websockets
asyncio
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart


